+++ test/integration/cartridge_hotreload_test.lua
@@ -1,14 +1,19 @@
 local fio = require('fio')
 local t = require('luatest')
-local g = t.group('cartridge-hotreload')
+local g = t.group()
+local utils = require('test.utils')
 
 local helpers = require('test.helper')
 
-g.before_each( function()
+g.before_each(function()
     t.skip_if(type(helpers) ~= 'table', 'Skip cartridge test')
+    local cartridge_version = require('cartridge.VERSION')
+    t.skip_if(cartridge_version == 'unknown')
+    t.skip_if(utils.is_version_less(cartridge_version, '2.3.0'))
     g.cluster = helpers.Cluster:new({
         datadir = fio.tempdir(),
         server_command = helpers.entrypoint('srv_basic'),
+        env = {TARANTOOL_ROLES_RELOAD_ALLOWED = 'true'},
         replicasets = {
             {
                 uuid = helpers.uuid('a'),
@@ -21,12 +26,12 @@ g.before_each( function()
         },
     })
     g.cluster:start()
-end )
+end)
 
-g.after_each( function()
+g.after_each(function()
     g.cluster:stop()
     fio.rmtree(g.cluster.datadir)
-end )
+end)
 
 local function upload_config()
     local main_server = g.cluster:server('main')
@@ -38,7 +43,7 @@ local function upload_config()
                     format = 'health'
                 },
                 {
-                    path = '/metrics',
+                    path = '/new-metrics',
                     format = 'json'
                 },
             },
@@ -46,45 +51,84 @@ local function upload_config()
     })
 end
 
-g.test_cartridge_hotreload = function()
-    local main_server = g.cluster:server('main')
-    t.skip_if(main_server.net_box:eval([[ return pcall(require, 'cartridge.roles') == false ]]))
-    t.skip_if(main_server.net_box:eval([[ return require('cartridge.roles').reload == nil ]]))
+local function set_export()
+    local export = {
+        {
+            path = '/health',
+            format = 'health'
+        },
+        {
+            path = '/metrics',
+            format = 'json'
+        },
+    }
+    local server = g.cluster.main_server
+    return server.net_box:eval([[
+        local cartridge = require('cartridge')
+        local metrics = cartridge.service_get('metrics')
+        local _, err = pcall(
+            metrics.set_export, ...
+        )
+        return err
+    ]], {export})
+end
 
+local function reload_roles()
+    local main_server = g.cluster:server('main')
     main_server.net_box:eval([[
-        box.schema.space.create('test'):create_index('pkey')
-        box.schema.space.create('vinni_test', {engine='vinyl'}):create_index('pkey')
-        box.space.test:put{1, 1}
-        box.space.vinni_test:put{1, 1}
+        require('cartridge.roles').reload()
     ]])
+end
+
+g.test_cartridge_hotreload_set_export = function()
+    local main_server = g.cluster:server('main')
+    local resp = main_server:http_request('get', '/metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+
+    set_export()
+
+    resp = main_server:http_request('get', '/metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+
+    reload_roles()
+
+    main_server = g.cluster:server('main')
+    resp = main_server:http_request('get', '/metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+end
+
+g.test_cartridge_hotreload_config = function()
+    local main_server = g.cluster:server('main')
 
     upload_config()
-    local resp = main_server:http_request('get', '/health')
+    local resp = main_server:http_request('get', '/new-metrics')
     t.assert_equals(resp.status, 200)
 
-    resp = main_server:http_request('get', '/metrics')
-    t.assert_str_contains(resp.body, '"test"')
-    t.assert_str_contains(resp.body, '"vinni_test"')
+    reload_roles()
 
     main_server = g.cluster:server('main')
-    main_server.net_box:eval([[
-        require('cartridge.roles').reload()
-    ]])
+    resp = main_server:http_request('get', '/new-metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+end
+
+g.test_cartridge_hotreload_set_export_and_config = function()
+    local main_server = g.cluster:server('main')
+
+    set_export()
+
+    upload_config()
+    local resp = main_server:http_request('get', '/new-metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+
+    resp = main_server:http_request('get', '/metrics', {raise = false})
+    t.assert_equals(resp.status, 200)
+
+    reload_roles()
 
     main_server = g.cluster:server('main')
-    resp = main_server:http_request('get', '/health')
+    resp = main_server:http_request('get', '/new-metrics', {raise = false})
     t.assert_equals(resp.status, 200)
 
-    local replica = g.cluster:server('replica')
-    replica.net_box:eval([[
-        require('cartridge.roles').reload()
-    ]])
-
-    replica = g.cluster:server('replica')
-    resp = replica:http_request('get', '/health')
+    resp = main_server:http_request('get', '/metrics', {raise = false})
     t.assert_equals(resp.status, 200)
-
-    resp = main_server:http_request('get', '/metrics')
-    t.assert_str_contains(resp.body, '"test"')
-    t.assert_str_contains(resp.body, '"vinni_test"')
 end